add admin navbar in themplate

+++ app/Models/Widget.php
@@ -78,9 +78,11 @@ class Widget extends Model
                 $public_destination_path = Str::replaceFirst('public/', '', $destination_path);
                 $values[$attribute] = $public_destination_path.$filename;
             }
-            elseif ($value == null) {
+            elseif ($value == null and $this->{$attribute} != null) {
                 // delete the image from disk
-                Storage::disk($disk)->delete($this->{$attribute});
+                if(Storage::disk($disk)->exists($this->{$attribute})) {
+                    Storage::disk($disk)->delete($this->{$attribute});
+                }
             }
         }
